Fix macroexpand1's handling of ArityException
macroexpand1 would simply throw a new ArityException in order to edit
the number of arguments, hiding the fact that the two extra arguments
beyond a macro's signature, &form and &env, are passed to a macro when
it's invoked.  This throw would erase any stack info below macroexpand1.
If the ArityException was actually caused by something a call internal
to the macro, this would result in very confusing stack traces.  All you
would get is the name of the function where the ArityException was seen,
and the number of arguments would be off by two.

This patch fixes this first by checking the number of non-java
stackframes below macroexpand1.  If there are any, then the
ArityException was not from the macro macroexpand1 called, so it leavs
the stacktrace alone.  Otherwise, it creates a new ArityException newexc
with the number of arguments reduced by two, and puts the original
stacktrace elements in newexc before throwing it.  This preserves the
location data for the ArityException.